feat(frontend): split ic-wallet.scheduler
