+++ src/frontend/src/icp/workers/icp-wallet.worker.ts
@@ -1,5 +1,5 @@
 import { getTransactions as getTransactionsApi } from '$icp/api/icp-index.api';
-import { IcWalletScheduler } from '$icp/schedulers/ic-wallet.scheduler';
+import { IcWalletTransactionsScheduler } from '$icp/schedulers/ic-wallet-transactions.scheduler';
 import type { IcTransactionAddOnsInfo, IcTransactionUi } from '$icp/types/ic-transaction';
 import { mapIcpTransaction, mapTransactionIcpToSelf } from '$icp/utils/icp-transactions.utils';
 import type { SchedulerJobData, SchedulerJobParams } from '$lib/schedulers/scheduler';
@@ -32,8 +32,16 @@ const mapTransaction = ({
 	jobData: SchedulerJobData<PostMessageDataRequest>;
 }): IcTransactionUi => mapIcpTransaction({ transaction, identity });
 
-const scheduler: IcWalletScheduler<Transaction, TransactionWithId, PostMessageDataRequest> =
-	new IcWalletScheduler(getTransactions, mapTransactionIcpToSelf, mapTransaction, 'syncIcpWallet');
+const scheduler: IcWalletTransactionsScheduler<
+	Transaction,
+	TransactionWithId,
+	PostMessageDataRequest
+> = new IcWalletTransactionsScheduler(
+	getTransactions,
+	mapTransactionIcpToSelf,
+	mapTransaction,
+	'syncIcpWallet'
+);
 
 onmessage = async ({ data: dataMsg }: MessageEvent<PostMessage<PostMessageDataRequest>>) => {
 	const { msg, data } = dataMsg;
